pyproject.toml
[project]
name = "img-gen-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.16.1",
    "mcp[cli]>=1.9.1",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
]
